✨ feat(book/frame): create new BookFramePreview component
- move frame preview functionality from BookFrame to BookFramePreview
- remove unused code related to frame preview in BookFrame
- import and use the BookFramePreview component in BookFrame
